README.md: added missing rules from RFC 2234
The "Core rules" section of this document was missing the final items from Section 6.1 of RFC 2234, "Core Rules".  They exist at the top of page 12 of that document (from the "; space" for SP (the core rule having been included, only missing the comment) through the WSP definition), and were simply missing in the README.  I've not checked to ensure that the code handles them correctly, though I presume it does.